Enabled Alignment Selector
Built another generator step and it seems to prove out the value of splitting out all this logic as it was simple to implement. Improved gateways with caching so it doesn't reload ever request. This will make performance better but it has a downside of ignoring file changes. Might need to revisit that at some point.