Completed Commoner class
This includes a change to support dynamic values. It's not smooth yet and will take some more to flush out but it does support changing what values are there without reworking the entire system which is good. I have a feeling this will be more of a base of all property values in order for anywhere to be generated dynamically.