Migrated appearance enums to data files
Now it will be easier to create special case logic for which styles and colors to pick. It also opens the possibility for more sophisticated descriptions by tying supporting adjectives to different items. For example, black eyes might be described as midnight black eyes. Or maybe blue is tropical surf or something cheesy. But the point is that it allows for more customization much more easily. Also, certain attributes could be limited by specific values so certain genders, races, etc...