Reduced redundancy and simplified code to add abilities
The logic that added feats, traits and abilities was pretty close to complete duplication. This has been somewhat simplified to provide a bit more consistency and also will open up the door for feats, abilities, and traits to all work more similiarly. In the long run I see a merging of these different pieces into a common piece of code