Updated to support visual studio and broken out solutions
This update changes how everything is layed out. This is to support the new versions of .net core and also to work with heroku again. This break out also will allow the library to be better support to be migrated to other systems such as Unity. The downside is that currently unit tests are not functional and will require more work to activate.