WeaponTraining migrated to ClassFeature style.
Required some refactoring that is a little dodgy and incomplete. Something has to be improved about modifiers so that they have flexibility but also ease of use. The multiple classes for different types is maybe not ideal, or there should be a base implementation that is maximimally flexible and the inherited classes hide that complexity.